Ignore swap files when staging
When editing a file using Vim, a (temporary) swap file is generated, and
removed when Vim is closed. However, if vim-fugitive is used to make
commits—or any other Git plugin for Vim—while editing, the swap files
are still present at staging time. In order to avoid version-controlling
them, they are added to the gitignore.