Create an Explorer class for data processing
The goal of the Explorer class is to provide an interface that allows to
quickly access useful functions (like plot and plot_files) that share
variables (especially the name converter). Class attributes allow to do
just that.
Additionally, basic thermodynamic calculations can also be implemented as
instance methods. An Explorer should also be able to pull out its data
to allow calculations that are not implemented.
For now, only the methods read and _create_name_converter have already
been written.