Add plot method and make build_quantities private
While writing the plot metod, it was found that in many cases it woud be
more efficient to add missing quantities to the Explorer instead of
updating them all. This feature was already implemented in the get
method, but this method should be used exclusively for retrieving
purposes.

Accordingly, an update option is added to the build_quantities method,
which is made private (hence renamed _build_quantities) for it doesn't
have to return anything anymore.