Plot magnitude to avoid UnitStrippedWarnings
When plotting a Quantity, pint takes the current unit by default (unit
specification in axis does not seem to work for now). Accordingly, a
'UnitStrippedWarning' is raised. To avoid displaying this message, the
magnitude of the quantity is passed to the matplotlib.pyplot.plot
function instead of the quantity itself.