Store timestamps as pandas Timestamp objects
In order to avoid an error when computing the time step from the first two timestamps,
the raw_data index is reset after concatenation. That way, raw_data['Timestamp'][0]
will return a single Timestamp object, instead of a Series of timestamps.

When a single path is provided, it is put in a list to avoid an error when iterating on paths.